empaquetado de comentario
